Remove link-atom scope from electrostatic embedding
