+++ doc/userguide.md
@@ -105,8 +105,6 @@ When using MACE models, the following extra keyword arguments to `createSystem()
 | --- | --- |
 | `precision` | The numerical precision of the model. Supported options are `'single'` and `'double'`.  If `None`, the default precision of the model is used. |
 | `returnEnergyType` | Whether to return the interaction energy or the energy including the self-energy.  The default is `'energy'`, which is the scalar the reported forces are the gradient of, so the potential is conservative.  Supported options are `'interaction_energy'` and `'energy'`. |
-| `linkRecords` | Hydrogen link atoms to cap covalent bonds crossing the ML/MM boundary; see *Link atoms* below.  Only meaningful for `createMixedSystem()`. |
-| `linkChargeScheme` | How to treat the MM charges on the MM-side boundary atoms when `linkRecords` is given; see *Link atoms* below.  The default is `'dz1'`. |
 | `embedding` | Passed by the embedding method rather than by you; see *Embeddings* below. |
 | `device` | The PyTorch device to perform calculations on, either a `torch.device` object or a string (such as `'cuda'` or `'cpu'`.)  If omitted, a device is chosen automatically. |
 | `charge` | The total charge of the system.  If omitted, it is assumed to be 0.  This is only used by MACE-OMOL-0.  For other models it is ignored. |
@@ -413,27 +411,6 @@ endpoint.
 ```python
 system = potential.createMixedSystem(topology, mm_system, ml_atoms, embedding='electrostatic')
 ```
-
-#### Link atoms
-
-When the ML subset is carved out of the middle of a molecule, the covalent bonds crossing the boundary leave the ML
-atoms with dangling valences.  The MACE interface can cap each such bond with a hydrogen link atom, placed along the
-bond vector at every force evaluation and with its forces redistributed onto the two real atoms it sits between, so the
-capped system stays conservative.  Caps are requested with the `linkRecords` argument to `createMixedSystem()`, which
-takes a sequence of `(q, m, target_dist)` tuples — `q` is the ML-side atom index, `m` the MM-side atom index, and
-`target_dist` the desired cap distance from `q` in Ångströms — or the path to a CSV file with `q_idx1`, `m_idx1`, and
-`target_dist_ang` columns.  Note that the tuples are 0-based, matching OpenMM's atom indices, while the CSV columns are
-1-based, hence their names.
-
-The MM partial charge on each MM-side boundary atom would otherwise sit about 1.5 Å from the nearest ML atom, through
-the cap, and over-polarise it.  The `linkChargeScheme` argument controls how that charge is handled, affecting only the
-charges passed to the MLIP; the MM `NonbondedForce` is left alone, so MM-MM electrostatics remain exact.
-
-| `linkChargeScheme` | Description |
-| --- | --- |
-| `'dz1'` | The default.  Sets the MM-side boundary atom charge to zero and spreads it evenly over that atom's MM neighbours, conserving total charge.  Falls back to `'z1'`, with a warning, for a boundary atom with no MM neighbours. |
-| `'z1'` | Sets the MM-side boundary atom charge to zero without redistributing it, which changes the total charge. |
-| `'none'` | Leaves the MM charges untouched. |
 
 ### Molecules Spanning the ML-MM Region
