+++ test/TestPolarMACEElectrostaticEmbedding.py
@@ -464,195 +464,3 @@ def test_polar_mace_electrostatic_mm_charge_displacement_changes_ml_force(
     # responds to the shifted MM field.
     delta_ml = f_b[:3] - f_a[:3]
     assert np.linalg.norm(delta_ml) > 1e-3
-
-
-# ---------------------------------------------------------------------------
-# Link-atom CHARGE redistribution (Z1 / DZ1) — end-to-end via MACE
-# ---------------------------------------------------------------------------
-#
-# The unit tests in TestLinkChargeRedistribution.py cover the pure helper
-# (apply_link_charge_redistribution). The tests below exercise the actual
-# integration into MACEPotentialImpl.addForces and verify that:
-#   (a) the scheme parameter propagates all the way to MACE's mm_charges
-#       input (different schemes give different energies/forces);
-#   (b) forces remain conservative under each scheme (F = -dE/dx, FD-tested).
-
-
-def _build_capped_nonpbc_chain():
-    """Tiny 6-atom chain: ML region {0,1,2} bonded to MM atom 3 (the M atom)
-    which is bonded to MM atoms {4,5} (M's neighbours, for DZ1 to populate).
-    All MM atoms carry non-zero partial charges so the linkChargeScheme has
-    something to redistribute.
-    """
-    system = openmm.System()
-    # Make all six atoms hydrogen (element 1) so MACE's atomic-number map
-    # exercised by the polar_mace_model_path fixture works (it knows H+O only).
-    for _ in range(6):
-        system.addParticle(1.0)
-    nb = openmm.NonbondedForce()
-    # ML atoms get zero charge (they'll be re-zeroed by createMixedSystem anyway,
-    # but make it explicit). MM atoms get something the scheme can act on.
-    nb.addParticle(0.0,  0.30, 0.20)   # 0  ML
-    nb.addParticle(0.0,  0.30, 0.20)   # 1  ML
-    nb.addParticle(0.0,  0.30, 0.20)   # 2  ML  (Q atom in the cut)
-    nb.addParticle(-0.40, 0.30, 0.20)  # 3  MM  (the M atom)
-    nb.addParticle(0.15, 0.30, 0.20)   # 4  MM  (M1 neighbour)
-    nb.addParticle(0.25, 0.30, 0.20)   # 5  MM  (M1 neighbour)
-    nb.setNonbondedMethod(openmm.NonbondedForce.NoCutoff)
-    system.addForce(nb)
-
-    topology = app.Topology()
-    chain = topology.addChain()
-    res = topology.addResidue("X", chain)
-    atoms = [topology.addAtom(f"H{i}", elem.hydrogen, res) for i in range(6)]
-    # Bonds: 0-1, 1-2, 2-3 (Q-M cut), 3-4 (M-M1), 3-5 (M-M1)
-    for a, b in [(0, 1), (1, 2), (2, 3), (3, 4), (3, 5)]:
-        topology.addBond(atoms[a], atoms[b])
-    return topology, system
-
-
-def _capped_chain_positions():
-    return np.array([
-        [0.00, 0.0, 0.0],
-        [0.11, 0.0, 0.0],
-        [0.22, 0.0, 0.0],
-        [0.33, 0.0, 0.0],   # M atom
-        [0.40, 0.07, 0.0],  # M1
-        [0.40, -0.07, 0.0], # M1
-    ]) * unit.nanometer
-
-
-def _build_capped_mixed(potential, scheme):
-    topology, mm_system = _build_capped_nonpbc_chain()
-    mixed = potential.createMixedSystem(
-        topology, mm_system, atoms=[0, 1, 2],
-        embedding="electrostatic",
-        linkRecords=[(2, 3, 1.09)],   # Q=2 (in ML), M=3 (MM)
-        linkChargeScheme=scheme,
-    )
-    plat = openmm.Platform.getPlatformByName("Reference")
-    ctx = openmm.Context(mixed, openmm.VerletIntegrator(0.001), plat)
-    ctx.setPositions(_capped_chain_positions())
-    return mixed, ctx
-
-
-# elem import for the topology builder; sits with the openmm.app namespace.
-from openmm.app import element as elem  # noqa: E402  (kept near point of use)
-
-
-def _mmChargesReachingModel(potential, scheme):
-    """The mm_charges array that createMixedSystem bakes into the ML force.
-
-    Asserting on the array rather than on the energy it produces is deliberate.
-    The energy route would only distinguish the schemes if the model in use
-    responded to the MM charge *values*, and the synthetic checkpoint built by
-    the polar_mace_model_path fixture does not for this fixture's geometry: its
-    energy is bit-identical whether mm_charges are passed unchanged, zeroed, or
-    scaled tenfold.  What the redistribution schemes are responsible for is the
-    contents of that array, so that is what these tests pin, exactly rather
-    than as an inequality.
-    """
-    captured = {}
-    original = macepotential._computeMACE
-
-    def capture(state, **args):
-        if args.get("mmInfo") is not None and "mm_charges" not in captured:
-            captured["mm_atoms"] = np.asarray(args["mmInfo"]["mm_atoms"])
-            captured["mm_charges"] = np.asarray(args["mmInfo"]["mm_charges"])
-        return original(state, **args)
-
-    macepotential._computeMACE = capture
-    try:
-        _, context = _build_capped_mixed(potential, scheme=scheme)
-        context.getState(getEnergy=True)
-    finally:
-        macepotential._computeMACE = original
-    assert "mm_charges" in captured, "the ML force never received MM charges"
-    return captured["mm_atoms"], captured["mm_charges"]
-
-
-def test_link_charge_scheme_none_passes_forcefield_charges(polar_mace_model_path):
-    """With scheme='none' the model sees the MM force field charges verbatim."""
-    potential = MLPotential("mace", modelPath=polar_mace_model_path)
-    mmAtoms, mmCharges = _mmChargesReachingModel(potential, "none")
-    # Atoms 3 (the M atom), 4 and 5 (its MM neighbours) from
-    # _build_capped_nonpbc_chain.
-    np.testing.assert_array_equal(mmAtoms, [3, 4, 5])
-    np.testing.assert_allclose(mmCharges, [-0.40, 0.15, 0.25], atol=1e-12)
-
-
-def test_link_charge_scheme_z1_zeros_the_m_atom(polar_mace_model_path):
-    """Z1 zeros the M-atom charge and leaves every other MM charge alone,
-    which changes the total MM charge by -q_M."""
-    potential = MLPotential("mace", modelPath=polar_mace_model_path)
-    _, mmCharges = _mmChargesReachingModel(potential, "z1")
-    np.testing.assert_allclose(mmCharges, [0.0, 0.15, 0.25], atol=1e-12)
-    assert mmCharges.sum() == pytest.approx(0.40, abs=1e-12)
-
-
-def test_link_charge_scheme_dz1_spreads_the_m_charge_and_conserves(polar_mace_model_path):
-    """DZ1 zeros the M-atom charge and spreads it over M's two MM neighbours,
-    -0.40/2 = -0.20 each, leaving the total MM charge unchanged."""
-    potential = MLPotential("mace", modelPath=polar_mace_model_path)
-    _, mmCharges = _mmChargesReachingModel(potential, "dz1")
-    np.testing.assert_allclose(mmCharges, [0.0, 0.15 - 0.20, 0.25 - 0.20], atol=1e-12)
-    assert mmCharges.sum() == pytest.approx(0.0, abs=1e-12)
-
-
-@pytest.mark.parametrize("scheme", ["none", "z1", "dz1"])
-def test_link_charge_scheme_fd_force_consistency(polar_mace_model_path, scheme):
-    """For each scheme, the MACE-side forces must equal the negative gradient
-    of the MACE-side energy. Central finite difference in f64.
-
-    Atoms 2 (Q) and 3 (M) are included deliberately: the cap sits between them
-    and is repositioned from their coordinates every step, so they are the only
-    atoms whose forces exercise the cap chain rule in redistribute_cap_force.
-    Perturbing only the interior QM atoms would leave that untested.
-
-    The criterion is relative to the size of the force being checked, not an
-    absolute number of kJ/mol/nm. This runs against a synthetic *untrained*
-    checkpoint at an arbitrary geometry, where forces reach ~1e7 kJ/mol/nm on
-    the boundary atoms and only ~1e6 in the interior; a fixed absolute
-    threshold is meaninglessly loose for the former and tight for the latter.
-    The central-difference truncation error scales as delta^2 while a genuine
-    force/energy inconsistency does not shrink with delta at all, which is what
-    separates the two and what this bound is sized against.
-    """
-    potential = MLPotential("mace", modelPath=polar_mace_model_path)
-    _, ctx = _build_capped_mixed(potential, scheme=scheme)
-
-    pos0 = _capped_chain_positions().value_in_unit(unit.nanometer).copy()
-    delta = 1e-5  # nm; the truncation term goes as delta^2
-
-    ctx.setPositions(pos0 * unit.nanometer)
-    e_ref = ctx.getState(getEnergy=True).getPotentialEnergy().value_in_unit(
-        unit.kilojoules_per_mole)
-    f_ref = ctx.getState(getForces=True).getForces(asNumpy=True).value_in_unit(
-        unit.kilojoules_per_mole / unit.nanometer)
-
-    max_abs_err = 0.0
-    # Atoms 2 (Q) and 3 (M) are the ones the cap is placed between, so they are
-    # the only ones whose forces exercise the cap chain rule. Perturbing just
-    # the interior QM atoms 0 and 1 would leave the redistribution untested.
-    for ia in (0, 1, 2, 3):
-        for ax in range(3):
-            saved = pos0[ia, ax]
-            pos0[ia, ax] = saved + delta
-            ctx.setPositions(pos0 * unit.nanometer)
-            e_p = ctx.getState(getEnergy=True).getPotentialEnergy().value_in_unit(
-                unit.kilojoules_per_mole)
-            pos0[ia, ax] = saved - delta
-            ctx.setPositions(pos0 * unit.nanometer)
-            e_m = ctx.getState(getEnergy=True).getPotentialEnergy().value_in_unit(
-                unit.kilojoules_per_mole)
-            pos0[ia, ax] = saved
-            f_num = -(e_p - e_m) / (2.0 * delta)
-            err = abs(f_num - f_ref[ia, ax])
-            max_abs_err = max(max_abs_err, err)
-    forceScale = float(np.abs(f_ref).max())
-    max_rel_err = max_abs_err / forceScale
-    assert max_rel_err < 5e-5, (
-        f"FD vs analytical force mismatch for scheme='{scheme}': "
-        f"max |dF| = {max_abs_err:.3e} kJ/mol/nm against a force scale of "
-        f"{forceScale:.3e}, i.e. relative {max_rel_err:.3e} (expected < 5e-5)"
-    )
